Refactor GDTF control resolver into focused DMX modules
